kun en behandlingform, nå typet (#210)
